Move profiling to another file
